Mark all transactions originating from this plugin as "not for history"
Changes coming through Yjs are not changes from the local user, but rather they represent
changes in other places in the system. Such changes do not need to be tracked and made
undo-able. This plugin also triggers prosemirror changes when rendering snapshots
and when forcing re-renders, which do a full document replace step.

By marking these with the `"addToHistory"` meta the prosemirror-history plugin will be
able to ignore these changes for the purposes of providing undo/redo functionality on the
editor model level, without affecting a Yjs document level undo/redo.